requirements.txt
# Django dependencies
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1

# Django Rest Framework dependencies
djangorestframework==3.12.4
importlib-metadata==3.9.0
typing-extensions==3.7.4.3
zipp==3.4.1

# API Pagination
django-filter==2.4.0

# Module for ImageField
Pillow==8.2.0

# JsonWebToken
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1

# CORS Management
django-cors-headers==3.7.0
 
# API Documentation
Markdown==3.3.4
PyYAML==5.4.1
uritemplate==3.0.1
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pyparsing==2.4.7
requests==2.25.1
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
urllib3==1.26.4
